requirements.txt
Django==3.1   # Django3.1
line-bot-sdk==1.8.0
google-cloud-vision==1.0.0